Add `WithService()` methods to gsheets & gdrive
This allows consumers of this (very nifty!) package to override the
service and provide their own (vs the default). Useful when using custom
options (ex specifying a credentials file or different transport).

An alternative would be to take the service as an argument to existing
constructor (breaking change, did not consider) or to add make the
internal interface public (not pursued since that interface is private
and I thought it odd to leak).

This change does **not** impact CLI usage and only expands the
functionality as a go package.